TcpServerChannel works

+++ source/msgtrans/MessageBuffer.d
@@ -4,10 +4,10 @@ import std.format;
 
 class MessageBuffer
 {
-    size_t id;
+    uint id;
     ubyte[] data;
 
-    this(size_t id, ubyte[] data) {
+    this(uint id, ubyte[] data) {
         this.id = id;
         this.data = data;
     }
